+++ src/layers/Google.js
@@ -204,7 +204,10 @@ L.Google.asyncInitialize = function() {
 	var el = document.createElement('script');
 	el.type = 'text/javascript';
 	el.src = 'http://maps.google.com/maps/api/js?v=3&sensor=false&callback=L.Google.asyncInitialize';
-	L.DomEvent.on(window, 'load', function() { document.body.appendChild(el); });
+	if( document.readyState == 'interactive' || document.readyState == 'complete' )
+		document.body.appendChild(el);
+	else
+		L.DomEvent.on(window, 'load', function() { document.body.appendChild(el); });
 })();
 
 if( !('layerList' in window) )